% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{act}
\alias{act}
\alias{act.MDPModel}
\alias{act.MDPSample}
\title{Perform an Action}
\usage{
act(model, state, action, fast = FALSE, ...)

\method{act}{MDPModel}(model, state, action = NULL, fast = FALSE, ...)

\method{act}{MDPSample}(model, state, action, fast = FALSE, ...)
}
\arguments{
\item{model}{an MDP model.}

\item{state}{the current state.}

\item{action}{the chosen action. If the action is not specified (\code{NULL}) and
the MDP model contains a policy, then the action is chosen according to the
policy.}

\item{fast}{logical; if \code{TRUE} then extra state id to label conversions are avoided.}

\item{...}{if action is unspecified, then the additional parameters are
passed on to \code{\link[=action]{action()}} to determine the action using the model's policy.}
}
\value{
a named list with
the \code{reward} and the next state \code{state_prime}.
}
\description{
Performs an action in a state and returns the new state and reward.
}
\examples{
data(Maze)

act(Maze, "s(1,3)", "right")

# solve the maze and then ask for actions using the policy
sol <- solve_MDP(Maze)
act(sol, "s(1,3)")

# make the policy in sol epsilon-soft and ask 10 times for the action
replicate(10, act(sol, "s(1,3)", epsilon = .2))
}
\seealso{
Other MDP:
\code{\link[=MDP]{MDP()}},
\code{\link[=absorbing_states]{absorbing_states()}},
\code{\link{action_state_helpers}},
\code{\link[=available_actions]{available_actions()}},
\code{\link[=find_reachable_states]{find_reachable_states()}},
\code{\link[=reachable_states]{reachable_states()}},
\code{\link[=sample_MDP]{sample_MDP()}},
\code{\link[=sample_MDP.MDPSample]{sample_MDP.MDPSample()}},
\code{\link{start}},
\code{\link[=transition_graph]{transition_graph()}},
\code{\link[=transition_matrix]{transition_matrix()}},
\code{\link[=unreachable_states]{unreachable_states()}}

Other MDPSample:
\code{\link[=MDPSample]{MDPSample()}},
\code{\link[=absorbing_states]{absorbing_states()}},
\code{\link{action_state_helpers}},
\code{\link[=reachable_states]{reachable_states()}},
\code{\link[=sample_MDP.MDPSample]{sample_MDP.MDPSample()}},
\code{\link[=solve_MDP_PG]{solve_MDP_PG()}},
\code{\link{start}}
}
\author{
Michael Hahsler
}
\concept{MDP}
\concept{MDPSample}
