% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_actions.R
\name{available_actions}
\alias{available_actions}
\title{Available Actions in a State}
\usage{
available_actions(model, state, neg_inf_reward = TRUE, stay_in_place = FALSE)
}
\arguments{
\item{model}{a \link{MDP} object.}

\item{state}{a character vector specifying the states.}

\item{neg_inf_reward}{logical; consider an action that produced \code{-Inf} reward to all
end states unavailable?}

\item{stay_in_place}{logical; consider an action that results in the same state
with a probability of 1 as unavailable. Note that this means that
absorbing states have no available action!}
}
\value{
a character vector with the available actions.

a vector with the available actions.
}
\description{
Determine the set of actions available in a state.
}
\details{
Unavailable actions are modeled as actions that have an immediate
reward of \code{-Inf} in the reward function. For a maze, also actions that
do not change the state can be considered unavailable.
}
\examples{
data(DynaMaze)
gw_plot(DynaMaze)

# The following actions are always available:
DynaMaze$actions

# only right and down is unavailable for s(1,1) because they
#   make the agent stay in place.
available_actions(DynaMaze, state = "s(1,1)", stay_in_place = TRUE)

# An action that leaves the grid currently is allowed but does not do
# anything.
act(DynaMaze, "s(1,1)", "up")
}
\seealso{
Other MDP:
\code{\link[=MDP]{MDP()}},
\code{\link[=absorbing_states]{absorbing_states()}},
\code{\link[=act]{act()}},
\code{\link{action_state_helpers}},
\code{\link[=find_reachable_states]{find_reachable_states()}},
\code{\link[=reachable_states]{reachable_states()}},
\code{\link[=sample_MDP]{sample_MDP()}},
\code{\link[=sample_MDP.MDPSample]{sample_MDP.MDPSample()}},
\code{\link{start}},
\code{\link[=transition_graph]{transition_graph()}},
\code{\link[=transition_matrix]{transition_matrix()}},
\code{\link[=unreachable_states]{unreachable_states()}}
}
\author{
Michael Hahsler
}
\concept{MDP}
