% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_colors.R
\name{colors}
\alias{colors}
\alias{colors_discrete}
\alias{colors_continuous}
\title{Default Colors for Visualization}
\usage{
colors_discrete(n, col = NULL)

colors_continuous(val, col = NULL)
}
\arguments{
\item{n}{number of states.}

\item{col}{custom color palette. \code{colors_discrete()} uses the first n colors.
\code{colors_continuous()} uses the given colors to calculate a palette
(see \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}}). The default is a blue-red color ramp.}

\item{val}{a vector with values to be translated to colors.}
}
\value{
\code{colors_discrete()} returns a color palette and
\code{colors_continuous()} returns the colors associated with the supplied values.
}
\description{
Default discrete and continuous colors used in
the package \code{markovDP}.
}
\examples{
colors_discrete(5)

colors_continuous(runif(10))
}
\seealso{
Other utilities:
\code{\link[=round_stochastic]{round_stochastic()}}
}
\concept{utilities}
