% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return.R
\name{expected_return}
\alias{expected_return}
\alias{expected_return.MDP}
\title{Calculate the Expected Return of a Policy}
\usage{
expected_return(model, ...)

\method{expected_return}{MDP}(model, start = NULL, method = "solution", ...)
}
\arguments{
\item{model}{a solved \link{MDP} object.}

\item{...}{further arguments are passed on to \code{\link[=policy_evaluation]{policy_evaluation()}} or \code{\link[=sample_MDP]{sample_MDP()}}.}

\item{start}{specification of the current state (see argument start
in \link{MDP} for details). By default the start state defined in
the model as start is used. Multiple states can be specified as rows in a matrix.}

\item{method}{\code{"solution"} uses the converged value function stored in the solved model,
\code{"policy_evaluation"} estimates the value function, and \code{"sample"}
calculates the average return by sampling episodes from the model.}
}
\value{
\code{expected_return()} returns a vector of returns, one for each start state if a matrix is specified.

\item{state}{start state to calculate the return for. If \code{NULL} then the start
state of model is used.}
}
\description{
This function calculates the expected total return for an MDP policy
given a start state (distribution). The value is calculated using the value
function stored in the MDP solution.
}
\details{
The return is typically calculated using the value function
of the solution. If these are not available, then \code{\link[=sample_MDP]{sample_MDP()}} is
used instead with a warning.
}
\examples{
data("Maze")
Maze
gw_matrix(Maze)

sol <- solve_MDP(Maze)
policy(sol)

# return for the start state s(3,1) specified in the model
expected_return(sol)

# return for starting next to the goal at s(1,3)
expected_return(sol, start = "s(1,3)")

# expected return when we start from a random state as returned from the solver
expected_return(sol, start = "uniform")

# estimate the return using sampling following the policy
expected_return(sol, method = "sample", start = "uniform", n = 10000, horizon = 1000)
}
\seealso{
Other policy:
\code{\link[=action]{action()}},
\code{\link[=greedy_action]{greedy_action()}},
\code{\link[=policy]{policy()}},
\code{\link[=policy_evaluation]{policy_evaluation()}},
\code{\link[=regret]{regret()}},
\code{\link[=visit_probability]{visit_probability()}}
}
\author{
Michael Hahsler
}
\concept{policy}
