% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\alias{start.MDPModel}
\alias{start.MDPSample}
\title{Sample a Start State}
\usage{
\method{start}{MDPModel}(x, as = "factor", ...)

\method{start}{MDPSample}(x, as = "features", ...)
}
\arguments{
\item{x}{a model.}

\item{as}{a returned format. See parameter \code{as} in \code{\link[=normalize_state]{normalize_state()}}.}

\item{...}{additional parameters are ignored.}
}
\value{
a start state.
}
\description{
Samples a start state which can be used for simulation.
}
\examples{
data(Maze)

start(Maze)
}
\seealso{
Other MDP: 
\code{\link{MDP}()},
\code{\link{Q_values}()},
\code{\link{absorbing_states}()},
\code{\link{act}()},
\code{\link{available_actions}()},
\code{\link{bellman_update}()},
\code{\link{convergence_horizon}()},
\code{\link{greedy_action}()},
\code{\link{gridworld}},
\code{\link{policy_evaluation}()},
\code{\link{reachable_states}()},
\code{\link{regret}()},
\code{\link{sample_MDP}()},
\code{\link{solve_MDP}()},
\code{\link{transition_graph}()},
\code{\link{transition_matrix}()},
\code{\link{unreachable_states}()},
\code{\link{value_function}()}

Other MDPSample: 
\code{\link{MDPSample}()},
\code{\link{absorbing_states}()},
\code{\link{act}()},
\code{\link{reachable_states}()},
\code{\link{sample_MDP.MDPSample}()},
\code{\link{solve_MDP}()},
\code{\link{solve_MDP_APPROX}()},
\code{\link{solve_MDP_PG}()}
}
\author{
Michael Hahsler
}
\concept{MDP}
\concept{MDPSample}
