% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unreachable_states.R
\name{unreachable_states}
\alias{unreachable_states}
\alias{remove_unreachable_states}
\title{Unreachable States}
\usage{
unreachable_states(
  model,
  horizon = Inf,
  sparse = "states",
  progress = TRUE,
  ...
)

remove_unreachable_states(model, ...)
}
\arguments{
\item{model}{a \link{MDP} object.}

\item{horizon}{only states that can be reached within the
horizon are reachable.}

\item{sparse}{logical; return a sparse logical vector?}

\item{progress}{logical; show a progress bar?}

\item{...}{further arguments are passed on.}
}
\value{
\code{unreachable_states()} returns a
logical vector indicating the unreachable states.

\code{remove_unreachable_states()} returns a model with all
unreachable states removed.
}
\description{
Find or removes unreachable states using the transition model.
}
\details{
The function \code{unreachable_states()} checks if states
cannot be reached from any other state. It performs a depth-first
search which can be slow.
The search breaks cycles to avoid an infinite loop.
The search depth can be restricted using
\code{horizon}.

The function \code{remove_unreachable_states()} simplifies a model by
removing unreachable states from the model description.
}
\examples{
# create a Maze with an unreachable state

maze_unreach <- gw_read_maze(
    textConnection(c("XXXXXX", 
                     "XS X X",
                     "X  XXX",
                     "X   GX",
                     "XXXXXX")))
gw_plot(maze_unreach)

unreachable_states(maze_unreach)
unreachable_states(maze_unreach, sparse = FALSE)

maze <- remove_unreachable_states(maze_unreach)
unreachable_states(maze)
gw_plot(maze)
}
\seealso{
Other MDP:
\code{\link[=MDP]{MDP()}},
\code{\link[=absorbing_states]{absorbing_states()}},
\code{\link[=act]{act()}},
\code{\link{action_state_helpers}},
\code{\link[=available_actions]{available_actions()}},
\code{\link[=find_reachable_states]{find_reachable_states()}},
\code{\link[=reachable_states]{reachable_states()}},
\code{\link[=sample_MDP]{sample_MDP()}},
\code{\link[=sample_MDP.MDPSample]{sample_MDP.MDPSample()}},
\code{\link{start}},
\code{\link[=transition_graph]{transition_graph()}},
\code{\link[=transition_matrix]{transition_matrix()}}
}
\author{
Michael Hahsler
}
\concept{MDP}
