% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_function.R
\name{value_function}
\alias{value_function}
\alias{plot_value_function}
\alias{V_zero}
\alias{V_random}
\title{Value Function}
\usage{
value_function(model, drop = TRUE)

plot_value_function(
  model,
  epoch = 1L,
  legend = TRUE,
  col = NULL,
  ylab = "Value",
  las = 3,
  main = NULL,
  ...
)

V_zero(model, value = 0)

V_random(model, min = 1e-06, max = 1)
}
\arguments{
\item{model}{a solved \link{MDP}.}

\item{drop}{logical; drop the list for converged, epoch-independent value functions.}

\item{epoch}{epoch for finite time horizon solutions.}

\item{legend}{logical; show legend.}

\item{col, ylab, las}{are passed on to \code{\link[graphics:barplot]{graphics::barplot()}}.}

\item{main}{a main title for the plot. Defaults to the name of the problem.}

\item{...}{further arguments are passed on to \code{\link[graphics:barplot]{graphics::barplot()}}`.}

\item{value}{value to initialize the value function with. Default is 0.}

\item{min, max}{minimum and maximum for the uniformly distributed random
state values.}
}
\value{
Returns the value function as a numeric vector with one value for each state or
as a matrix with rows for states and columns for epochs.
}
\description{
Extracts the value function from a solved MDP.
}
\examples{
data("Maze")
sol <- solve_MDP(Maze)
sol

value_function(sol)
plot_value_function(sol)

## finite-horizon problem
sol <- solve_MDP(Maze, horizon = 3)
policy(sol)
value_function(sol)
plot_value_function(sol, epoch = 1)
plot_value_function(sol, epoch = 2)
plot_value_function(sol, epoch = 3)

# For a gridworld we can also plot is like this
gw_plot(sol, epoch = 1)
gw_plot(sol, epoch = 2)
gw_plot(sol, epoch = 3)
}
\seealso{
Other value_function:
\code{\link[=Q_values]{Q_values()}},
\code{\link[=bellman_update]{bellman_update()}}
}
\author{
Michael Hahsler
}
\concept{value_function}
\keyword{hplot}
