% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{act}
\alias{act}
\title{Perform an Action}
\usage{
act(model, state, action)
}
\arguments{
\item{model}{a MDP model.}

\item{state}{the current state.}

\item{action}{the chosen action.}
}
\value{
a names list with the next \code{state} and the \code{reward}.
}
\description{
Performs an action in a state and returns the new state and reward.
}
\examples{
data(Maze)

act(Maze, "s(1,3)", "right")
}
\seealso{
Other MDP: 
\code{\link{MDP}()},
\code{\link{accessors}},
\code{\link{actions}()},
\code{\link{add_policy}()},
\code{\link{gridworld}},
\code{\link{policy_evaluation}()},
\code{\link{q_values}()},
\code{\link{reachable_and_absorbing}},
\code{\link{regret}()},
\code{\link{simulate_MDP}()},
\code{\link{solve_MDP}()},
\code{\link{transition_graph}()},
\code{\link{value_function}()}
}
\author{
Michael Hahsler
}
\concept{MDP}
