% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/act.R
\name{act}
\alias{act}
\title{Perform an Action}
\usage{
act(model, state, action = NULL, ...)
}
\arguments{
\item{model}{an MDP model.}

\item{state}{the current state.}

\item{action}{the chosen action. If the action is not specified (\code{NULL}) and
the MDP model contains a policy, then the action is chosen according to the
policy.}

\item{...}{if action is unspecified, then the additional parameters are
passed on to \code{action()} to determine the action.}
}
\value{
a names list with the \code{old_state}, the \code{action},
the next \code{reward} and \code{state}.
}
\description{
Performs an action in a state and returns the new state and reward.
}
\examples{
data(Maze)

act(Maze, "s(1,3)", "right")

# solve the maze and then ask for actions using the policy
sol <- solve_MDP(Maze)
act(sol, "s(1,3)")

# make the policy in sol epsilon-soft and ask 10 times for the action 
replicate(10, act(sol, "s(1,3)", epsilon = .2))
}
\seealso{
Other MDP: 
\code{\link{MDP}()},
\code{\link{accessors}},
\code{\link{actions}()},
\code{\link{add_policy}()},
\code{\link{gridworld}},
\code{\link{policy_evaluation}()},
\code{\link{q_values}()},
\code{\link{regret}()},
\code{\link{sample_MDP}()},
\code{\link{solve_MDP}()},
\code{\link{transition_graph}()},
\code{\link{unreachable_and_absorbing}},
\code{\link{value_function}()}
}
\author{
Michael Hahsler
}
\concept{MDP}
