% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{actions}
\alias{actions}
\title{Available Actions in a State}
\usage{
actions(x, state)
}
\arguments{
\item{x}{a \link{MDP} object.}

\item{state}{a character vector of length one specifying the state.}
}
\value{
a character vector with the available actions.

a vector with the available actions.
}
\description{
Determine the set of actions available in a state.
}
\details{
Unavailable actions are modeled here a actions that have an immediate
reward of \code{-Inf} in the reward function.
}
\examples{
data(Maze)
gridworld_matrix(Maze)

# The the following actions are always available:
Maze$actions

# available actions in 
actions(Maze, state = "s(3,1)")
}
\seealso{
Other MDP: 
\code{\link{MDP}()},
\code{\link{MDP_policy_functions}},
\code{\link{accessors}},
\code{\link{add_policy}()},
\code{\link{gridworld}},
\code{\link{reachable_and_absorbing}},
\code{\link{regret}()},
\code{\link{simulate_MDP}()},
\code{\link{solve_MDP}()},
\code{\link{transition_graph}()},
\code{\link{value_function}()}
}
\author{
Michael Hahsler
}
\concept{MDP}
