% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_package.R
\docType{package}
\name{markovDP-package}
\alias{markovDP}
\alias{markovDP-package}
\title{markovDP: Infrastructure for Discrete-Time Markov Decision Processes (MDP)}
\description{
Provides the infrastructure to define and analyze the solutions of discete-time Markov Decision Process (POMDP) models. Interfaces for various exact and approximate solution algorithms are available including value iteration, point-based value iteration and SARSOP. Smallwood and Sondik (1973) \doi{10.1287/opre.21.5.1071}.
}
\section{Key functions}{

\itemize{
\item Problem specification: \link{MDP}
\item Solvers: \code{\link[=solve_MDP]{solve_MDP()}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mhahsler/mdp}
  \item Report bugs at \url{https://github.com/mhahsler/mdp/issues}
}

}
\author{
\strong{Maintainer}: Michael Hahsler \email{mhahsler@lyle.smu.edu} (\href{https://orcid.org/0000-0003-2716-1405}{ORCID}) [copyright holder]

}
\keyword{internal}
