% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDPSample.R
\name{MDPSample}
\alias{MDPSample}
\title{Define an MDP With Only Sample Access}
\usage{
MDPSample(
  actions,
  transition_model,
  start,
  states = NULL,
  absorbing_states = NULL,
  discount = 0.9,
  horizon = Inf,
  info = NULL,
  name = NA
)
}
\arguments{
\item{actions}{a character vector specifying the names of the available
actions.}

\item{transition_model}{A transition function receiving the current state features and
returning the reward and the next state features.}

\item{start}{Specifies in which state the MDP starts.}

\item{states}{Optional state labels.}

\item{absorbing_states}{a single state or a list with absorbing states.}

\item{discount}{numeric; discount rate between 0 and 1.}

\item{horizon}{numeric; Number of epochs.}

\item{info}{A list with additional information.}

\item{name}{a string to identify the MDP problem.}
}
\value{
The function returns an object of class MDPSample which is list with
the model specification.
}
\description{
Defines a discrete-time Markov decision process with only sample access
as a transition
function between states using a factored state representation. The state space
may be continuous and infinite.
}
\details{
Defines a Markov decision process as a transition
function between state represented as feature vectors
(i.e., using a factored state representation). The state space may be
\itemize{
\item finite and discrete and specified in the problem,
\item finite and discrete but not specified in the problem,
\item continuous (implying infinite and not specified in the problem).
}

Compared to the specification in class \link{MDP}, no transition probabilities are
available. Instead of the reward function, reward is only returned from
calls to the transition function via \code{\link[=act]{act()}}. The state space may be specified.

To represent states, a factored state representation as a \strong{row} vector for a single
state or a matrix with row vectors for a set of states are used. State labels
can be constructed in the form \code{s(feature1, feature2, ...)}. Conversion between the factored
representation and state labels is available in \code{\link[=state2features]{state2features()}}
and \code{\link[=features2state]{features2state()}}. Since the state set is not directly represented,
\strong{state ids cannot be used!}

Reinforcement learning algorithms with approximation can be used to solve
these problems. See: \code{\link[=solve_MDP_APPROX]{solve_MDP_APPROX()}}.
}
\examples{
# Example 1: Define a simple 5x5 maze without walls

transition_model <- function(model, action, state) {
  # make sure action and state are in the correct format
  action <- normalize_action_label(action, model)
  state <- normalize_state_features(state, model)
  
  if (all(state == s(5, 5)))
    return(list(reward = 0, state_prime = state))

  sp <- state + switch(action,
    up =   c( -1, 0),
    down = c( +1, 0),
    left = c(  0,-1),
    right = c( 0,+1)
  )
  r <- -1

  # check bounds
  if (any(sp < 1) || any(sp > 5)) {
      sp <- state
   }

  # goal
  if (all(sp == s(5, 5)))
    r <- 100

  return(list(reward = r, state_prime = sp))
}

m <- MDPSample(actions = c("up", "right", "down", "left"),
          transition_model,
          start = s(1,1),
          absorbing_states = s(5, 5),
          name = "5x5 Maze")
m

act(m, s(1,1), "down")

# Reach the goal: reward = 100
act(m, s(5,4), "right")

# Illegal action: no movement and reward = -1
act(m, s(1,1), "up")

# Absorbing state: no movement and 0 reward
act(m, s(5,5), "up")


# Example 2: Solve using Linear Feature Approximation

# Note that we have not specified the state space
m
S(m)

# To use value function approximation, we need to specify the minimum
# and maximum of each state feature for scaling them to [0,1]. The rows
# and column indices used are between 0 and 5.
set.seed(1000)
sol <- solve_MDP_APPROX(m, horizon = 1000, n = 100,
          alpha = 0.01, epsilon = 0.55, 
          transformation = transformation_linear_basis, 
          min = c(0, 0), max = c(5, 5),
          verbose = FALSE)

# we can calculate the approximate q-values, but have to specify the states.
approx_Q_value(sol, state = rbind(s(1,2), s(4,5)))

# Note: since there is no explicit state space, we also do not have a 
#       precalculated policy! We can get approx. greedy actions.

approx_greedy_action(sol, state = s(4,5))
 
# Example 3: MDPSample with a specified state space  

# The same maze as above can be created with this gridworld helper.
# It specifies the finite state space.
m <- gw_maze_MDP(dim = c(5,5), start = "s(1,1)", goal = "s(5,5)", access = "sample")
m
S(m)

# With the defined state space, we do not need to supply min and max 
# for the transformation function.
set.seed(1000)
sol <- solve_MDP_APPROX(m, horizon = 1000, n = 100,
          alpha = 0.01, epsilon = 0.55, verbose = FALSE)

# we can use several gridworld helper
gw_plot(sol)
gw_matrix(sol, what = "value")

# and get a policy table
policy(sol)
}
\seealso{
Other MDPSample:
\code{\link[=absorbing_states]{absorbing_states()}},
\code{\link[=act]{act()}},
\code{\link{action_state_helpers}},
\code{\link[=reachable_states]{reachable_states()}},
\code{\link[=sample_MDP.MDPSample]{sample_MDP.MDPSample()}},
\code{\link[=solve_MDP_PG]{solve_MDP_PG()}},
\code{\link{start}}
}
\author{
Michael Hahsler
}
\concept{MDPSample}
