% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\alias{start.MDPModel}
\alias{start.MDPSample}
\title{Sample a Start State}
\usage{
\method{start}{MDPModel}(x, as = "factor", ...)

\method{start}{MDPSample}(x, as = "features", ...)
}
\arguments{
\item{x}{a model.}

\item{as}{a returned format. See parameter \code{as} in \code{\link[=normalize_state]{normalize_state()}}.}

\item{...}{additional parameters are ignored.}
}
\value{
a start state.
}
\description{
Samples a start state which can be used for simulation.
}
\examples{
data(Maze)

start(Maze)
}
\seealso{
Other MDP:
\code{\link[=MDP]{MDP()}},
\code{\link[=absorbing_states]{absorbing_states()}},
\code{\link[=act]{act()}},
\code{\link{action_state_helpers}},
\code{\link[=available_actions]{available_actions()}},
\code{\link[=find_reachable_states]{find_reachable_states()}},
\code{\link[=reachable_states]{reachable_states()}},
\code{\link[=sample_MDP]{sample_MDP()}},
\code{\link[=sample_MDP.MDPSample]{sample_MDP.MDPSample()}},
\code{\link[=transition_graph]{transition_graph()}},
\code{\link[=transition_matrix]{transition_matrix()}},
\code{\link[=unreachable_states]{unreachable_states()}}

Other MDPSample:
\code{\link[=MDPSample]{MDPSample()}},
\code{\link[=absorbing_states]{absorbing_states()}},
\code{\link[=act]{act()}},
\code{\link{action_state_helpers}},
\code{\link[=reachable_states]{reachable_states()}},
\code{\link[=sample_MDP.MDPSample]{sample_MDP.MDPSample()}},
\code{\link[=solve_MDP_PG]{solve_MDP_PG()}}
}
\author{
Michael Hahsler
}
\concept{MDP}
\concept{MDPSample}
