% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_graph.R
\name{transition_graph}
\alias{transition_graph}
\alias{plot_transition_graph}
\alias{curve_multiple_directed}
\title{Transition Graph}
\usage{
transition_graph(
  x,
  action = NULL,
  state_col = NULL,
  simplify_transitions = TRUE,
  remove_unavailable_actions = TRUE
)

plot_transition_graph(
  x,
  action = NULL,
  state_col = NULL,
  simplify_transitions = TRUE,
  main = NULL,
  ...
)

curve_multiple_directed(graph, start = 0.3)
}
\arguments{
\item{x}{object of class \link{MDP}.}

\item{action}{the name or id of an action or a set of actions. By default the transition model for all actions is returned.}

\item{state_col}{colors used to represent the states.}

\item{simplify_transitions}{logical; combine parallel transition arcs into a single arc.}

\item{remove_unavailable_actions}{logical; don't show arrows for unavailable actions.}

\item{main}{a main title for the plot.}

\item{...}{further arguments are passed on to \code{igraph::plot.igraph()}.}

\item{graph}{The input graph.}

\item{start}{The curvature at the two extreme edges.}
}
\value{
returns the transition model as an igraph object.
}
\description{
Returns the transition model as an \pkg{igraph} object.
}
\details{
The transition model of an MDP is a Markov Chain. This function extracts the transition model as
an igraph object.
}
\examples{
data("Maze")

g <- transition_graph(Maze)
g

plot_transition_graph(Maze)
plot_transition_graph(Maze,
  vertex.size = 20,
  edge.label.cex = .1, edge.arrow.size = .5, margin = .5
)

## Plot using the igraph library
library(igraph)
plot(g)

# plot with a different layout
plot(g,
  layout = igraph::layout_with_sugiyama,
  vertex.size = 20,
  edge.label.cex = .6
)

## Use visNetwork (if installed)
if (require(visNetwork)) {
  g_vn <- toVisNetworkData(g)
  nodes <- g_vn$nodes
  edges <- g_vn$edges

  visNetwork(nodes, edges) \%>\%
    visNodes(physics = FALSE) \%>\%
    visEdges(smooth = list(type = "curvedCW", roundness = .6), arrows = "to")
}
}
\seealso{
Other MDP: 
\code{\link{MDP}()},
\code{\link{accessors}},
\code{\link{act}()},
\code{\link{actions}},
\code{\link{add_policy}()},
\code{\link{bellman_update}()},
\code{\link{gridworld}},
\code{\link{policy_evaluation}()},
\code{\link{q_values}()},
\code{\link{regret}()},
\code{\link{sample_MDP}()},
\code{\link{solve_MDP}()},
\code{\link{unreachable_and_absorbing}},
\code{\link{value_function}()}
}
\concept{MDP}
