% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reachable_and_absorbing.R
\name{unreachable_and_absorbing}
\alias{unreachable_and_absorbing}
\alias{unreachable_states}
\alias{absorbing_states}
\alias{remove_unreachable_states}
\title{Unreachable and Absorbing States}
\usage{
unreachable_states(model, direct = FALSE, sparse = FALSE, ...)

absorbing_states(model, sparse = FALSE, ...)

remove_unreachable_states(model, direct = FALSE)
}
\arguments{
\item{model}{a \link{MDP} object.}

\item{direct}{logical; only consider if a state is reachable from
a different state (not necessarily from a start state). This is
much faster and useful for gridworlds.}

\item{sparse}{logical; return a sparse logical vector.}

\item{...}{further arguments are passed on.}
}
\value{
\code{unreachable_states()} returns a
logical vector indicating the unreachable states.

\code{absorbing_states()} returns a logical vector indicating
if the states are absorbing (terminal).

\code{remove_unreachable_states()} returns a model with all
unreachable states removed.
}
\description{
Find unreachable and absorbing states using the transition model.
}
\details{
The function \code{unreachable_states()} checks if states
cannot be reached from any other state.

The function \code{remove_unreachable_states()} simplifies a model by
removing unreachable states from the model description.

The function \code{absorbing_states()} checks if a state or a set of states are
absorbing (terminal states). A state is absorbing
if there is for all actions a probability of 1 for staying in the state.
}
\examples{
data(Maze)

gridworld_matrix(Maze)
gridworld_matrix(Maze, what = "labels")

# -1 and +1 are absorbing states
absorbing_states(Maze)
absorbing_states(Maze, sparse = TRUE)
which(absorbing_states(Maze))

# all states in the model are reachable
unreachable_states(Maze)
unreachable_states(Maze, sparse = TRUE)
which(unreachable_states(Maze))
}
\seealso{
Other MDP: 
\code{\link{MDP}()},
\code{\link{accessors}},
\code{\link{act}()},
\code{\link{actions}},
\code{\link{add_policy}()},
\code{\link{gridworld}},
\code{\link{policy_evaluation}()},
\code{\link{q_values}()},
\code{\link{regret}()},
\code{\link{sample_MDP}()},
\code{\link{solve_MDP}()},
\code{\link{transition_graph}()},
\code{\link{value_function}()}
}
\author{
Michael Hahsler
}
\concept{MDP}
