% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_function.R
\name{value_function}
\alias{value_function}
\alias{plot_value_function}
\title{Value Function}
\usage{
value_function(model, drop = TRUE)

plot_value_function(
  model,
  epoch = 1,
  legend = TRUE,
  col = NULL,
  ylab = "Value",
  las = 3,
  main = NULL,
  ...
)
}
\arguments{
\item{model}{a solved \link{MDP}.}

\item{drop}{logical; drop the list for converged, epoch-independent value functions.}

\item{epoch}{epoch for finite time horizon solutions.}

\item{legend}{logical; show legend.}

\item{col, ylab, las}{are passed on to \code{\link[graphics:barplot]{graphics::barplot()}}.}

\item{main}{a main title for the plot. Defaults to the name of the problem.}

\item{...}{further arguments are passed on to \code{\link[graphics:barplot]{graphics::barplot()}}`.}
}
\value{
the function as a numeric vector with one value for each state.
}
\description{
Extracts the value function from a solved MDP.
}
\examples{
data("Maze")
sol <- solve_MDP(Maze)
sol

value_function(sol)
plot_value_function(sol)

## finite-horizon problem
sol <- solve_MDP(Maze, horizon = 3)
policy(sol)
value_function(sol)
plot_value_function(sol, epoch = 1)
plot_value_function(sol, epoch = 2)
plot_value_function(sol, epoch = 3)

# For a gridworld we can also plot is like this
gridworld_plot(sol, epoch = 1)
gridworld_plot(sol, epoch = 2)
gridworld_plot(sol, epoch = 3)
}
\seealso{
Other policy: 
\code{\link{action}()},
\code{\link{policy}()},
\code{\link{policy_evaluation}()},
\code{\link{q_values}()},
\code{\link{reward}()}

Other MDP: 
\code{\link{MDP}()},
\code{\link{accessors}},
\code{\link{act}()},
\code{\link{actions}()},
\code{\link{add_policy}()},
\code{\link{gridworld}},
\code{\link{policy_evaluation}()},
\code{\link{q_values}()},
\code{\link{reachable_and_absorbing}},
\code{\link{regret}()},
\code{\link{simulate_MDP}()},
\code{\link{solve_MDP}()},
\code{\link{transition_graph}()}
}
\author{
Michael Hahsler
}
\concept{MDP}
\concept{policy}
\keyword{hplot}
